+++ GGJ/Assets/Waves.Generator/instances/Wave_001.cs
@@ -7,9 +7,9 @@ namespace Assets.Waves.Generator.instances
 {
     public class Wave_001 : Assets.Waves.Generator.Base
     {
-        public override double InternalFunction(double x)
+        public override float InternalFunction(float x)
         {
-            return -1.0*Math.Sin(8 * x) * Math.Sin(x - 1);
+            return (float) (-1.0*Math.Sin(8.0f * x) * Math.Sin(x - 1.0));
         }
     }
 }
